markdown-indexer/requirements.txt
markdown
transformers
faiss-cpu
azure-search-documents
markdown2
numpy
pandas